Make player a singular class
