--- lib/theme_park/blackjack/players.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'players/ai'
-require_relative 'players/dealer'
-require_relative 'players/user'
-
-module ThemePark
-  module Blackjack
-    module Players
-    end
-  end
-end
