--- lib/theme_park/blackjack/players/ai.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'player'
-
-module ThemePark
-  module Blackjack
-    module Players
-      class AI < Player
-      end
-    end
-  end
-end
