--- lib/theme_park/blackjack/players/dealer.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module ThemePark
-  module Blackjack
-    module Players
-      class Dealer < Player
-      end
-    end
-  end
-end
